+++ foxes/core/turbine.py
@@ -31,6 +31,8 @@ class Turbine:
         settings if given
     wind_farm_name: str, optional
         The name of the wind farm the turbine belongs to
+    cluster_name: str, optional
+        The name of the cluster the wind farm belongs to
 
     :group: core
 
@@ -46,6 +48,7 @@ class Turbine:
         D=None,
         H=None,
         wind_farm_name=None,
+        cluster_name=None,
     ):
         """
         Constructor.
@@ -72,6 +75,8 @@ class Turbine:
             settings if given
         wind_farm_name: str, optional
             The name of the wind farm the turbine belongs to
+        cluster_name: str, optional
+            The name of the cluster the wind farm belongs to
 
         """
         self.index = index
@@ -81,6 +86,7 @@ class Turbine:
         self.D = D
         self.H = H
         self.wind_farm_name = wind_farm_name
+        self.cluster_name = cluster_name
 
         self.mstates_sel = models_state_sel
         if self.mstates_sel is None: